+++ CHANGELOG.md
@@ -17,6 +17,7 @@ changes. Best viewed [here](https://google-grain.readthedocs.io/en/latest/change
 
 * Bug fixes:
   * Fixed bug in DataLoader where sharding remainder was dropped even when ShardOptions.drop_remainder=False.
+  * Fixes reference cycle in BatchMapDataset.
 
 ## Grain 0.2.18 (June 17, 2026)